requirements.txt
alembic==1.7.3
autopep8==1.5.7
bcrypt==3.2.0
blinker==1.4
boto3==1.19.8
botocore==1.22.8
cffi==1.14.6
click==8.0.1
colorama==0.4.4
coverage==5.5
cryptography==35.0.0
dnspython==2.1.0
email-validator==1.1.3
Flask==1.0.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==2.4.0
Flask-User==1.0.1.5
Flask-WTF==0.14.2
greenlet==1.1.2
idna==3.2
importlib-resources==5.2.2
inflect==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.6.1
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
Werkzeug==0.16.0
WTForms==2.3.3
zipp==3.6.0
